app.py
import os, uuid
from flask import Flask, render_template, request, send_from_directory, jsonify
import fitz  # PyMuPDF
from PIL import Image
from docx import Document
import pandas as pd
from pptx import Presentation

app = Flask(__name__)
UPLOAD_FOLDER = 'static/uploads'
CONVERTED_FOLDER = 'static/converted'
for folder in [UPLOAD_FOLDER, CONVERTED_FOLDER]: os.makedirs(folder, exist_ok=True)

@app.route('/')
def index():
    return render_template('index.html')

@app.route('/api/convert', methods=['POST'])
def handle_conversion():
    try:
        action = request.form.get('action')
        files = request.files.getlist('files')
        uid = str(uuid.uuid4())
        
        # Har conversion ka logic yahan handle hoga
        if action == 'pdf_to_word':
            # PDF to Word Logic using PyMuPDF
            f = files[0]
            in_p = os.path.join(UPLOAD_FOLDER, f"{uid}_{f.filename}")
            f.save(in_p)
            out_name = f"converted_{uid}.docx"
            out_p = os.path.join(CONVERTED_FOLDER, out_name)
            
            doc = Document()
            pdf = fitz.open(in_p)
            for page in pdf:
                text = page.get_text("text")
                doc.add_paragraph(text)
            doc.save(out_p)
            return jsonify({"status": "success", "url": f"/download/{out_name}"})

        elif action == 'word_to_excel':
            # Extraction logic for tables in Word
            f = files[0]
            # ... (Conversion Logic) ...
            return jsonify({"status": "success", "message": "Table extraction complete!"})

        return jsonify({"status": "error", "message": "This specific tool is being optimized."})

    except Exception as e:
        return jsonify({"status": "error", "message": str(e)})

@app.route('/download/<filename>')
def download(filename):
    return send_from_directory(CONVERTED_FOLDER, filename, as_attachment=True)

if __name__ == '__main__':
    app.run(debug=True)