requirements.txt
Flask==3.0.0
PyMuPDF==1.23.5
python-docx==1.1.0
pandas==2.1.3
python-pptx==0.6.21
Pillow==10.1.0
nbformat==5.9.2
openpyxl==3.1.2
werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0